# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ScoredPaper {
  paperId: string;
  title: string;
  /** @nullable */
  abstract: string | null;
  /** @nullable */
  year: number | null;
  /** @nullable */
  url: string | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  relevanceScore: number;
  rationale: string;
  /**
     * @minimum -1
     * @maximum 1
     * @nullable
     */
  feedback: number | null;
}

export type PaperFeedbackInputThumbsUpDown = typeof PaperFeedbackInputThumbsUpDown[keyof typeof PaperFeedbackInputThumbsUpDown];


export const PaperFeedbackInputThumbsUpDown = {
  NUMBER_MINUS_1: -1,
  NUMBER_1: 1,
} as const;

export interface PaperFeedbackInput {
  /** @minLength 1 */
  paperId: string;
  thumbsUpDown: PaperFeedbackInputThumbsUpDown;
}

export type PaperFeedbackThumbsUpDown = typeof PaperFeedbackThumbsUpDown[keyof typeof PaperFeedbackThumbsUpDown];


export const PaperFeedbackThumbsUpDown = {
  NUMBER_MINUS_1: -1,
  NUMBER_1: 1,
} as const;

export interface PaperFeedback {
  paperId: string;
  thumbsUpDown: PaperFeedbackThumbsUpDown;
  message: string;
}

export interface ApiError {
  error: string;
}