# lib/api-zod/src/generated/types/paperFeedback.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaperFeedbackThumbsUpDown } from './paperFeedbackThumbsUpDown';

export interface PaperFeedback {
  paperId: string;
  thumbsUpDown: PaperFeedbackThumbsUpDown;
  message: string;
}