# lib/api-zod/src/generated/types/paperFeedbackInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaperFeedbackInputThumbsUpDown } from './paperFeedbackInputThumbsUpDown';

export interface PaperFeedbackInput {
  /** @minLength 1 */
  paperId: string;
  thumbsUpDown: PaperFeedbackInputThumbsUpDown;
}