# lib/api-zod/src/generated/types/scoredPaper.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScoredPaper {
  paperId: string;
  title: string;
  /** @nullable */
  abstract: string | null;
  /** @nullable */
  year: number | null;
  /** @nullable */
  url: string | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  relevanceScore: number;
  rationale: string;
  /**
     * @minimum -1
     * @maximum 1
     * @nullable
     */
  feedback: number | null;
}